+++ src/pages/Footer.jsx
@@ -1,4 +1,4 @@
-import "./Footer.css";
+import "../styles/Footer.css";
 import {
   FaFacebook,
   FaGithub,
